a function that returns the number of element

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -13,7 +13,7 @@
 size_t listint_len(const listint_t *h)
 {
 const listint_t *m;
-size_t s = 0;
+int s = 0;
 if (h == NULL)
 {
 	return (0);
